v1.4.0 #2 New commands: Nudge selected item(s) +/- 1 sample Color selected item(s)/track(s) same random color Color selected item(s)/track(s) custom color... Added "Add action marker" ! command to ReaConsole Cleaned up clickable column handling in sws_wnd and made work OSX, too Made action markers work with ID strings as well as ID #s Renamed "export tracklist" to "export formatted marker list" Made OSX copy/paste marker set work ListView refresh performance enhancements

+++ ObjectState/TrackEnvelope.cpp
@@ -111,7 +111,7 @@ void SWS_TrackEnvelope::SetHeight(int iHeight)
 	lstrcpyn(m_cEnv, cOldState, pLH-cOldState + 12);
 	sprintf(m_cEnv+strlen(m_cEnv), "%d", iHeight);
 	strcpy(m_cEnv+strlen(m_cEnv), strchr(pLH+11, ' '));
-	GetSetEnvelopeState(m_pTe, m_cEnv, strlen(m_cEnv));
+	GetSetEnvelopeState(m_pTe, m_cEnv, (int)strlen(m_cEnv));
 	free(cOldState);
 }
 
@@ -134,7 +134,7 @@ void SWS_TrackEnvelope::SetVis(bool bVis)
 		return;
 
 	pLH[4] = bVis ? '1' : '0';
-	GetSetEnvelopeState(m_pTe, m_cEnv, strlen(m_cEnv));
+	GetSetEnvelopeState(m_pTe, m_cEnv, (int)strlen(m_cEnv));
 }
 
 SWS_TrackEnvelopes::SWS_TrackEnvelopes():m_pTr(NULL)
